Rebase profile 5

+++ app/src/main/java/ru/itis/tinkoff/project/domain/repositories/UserRepository.kt
@@ -3,5 +3,5 @@ package ru.itis.tinkoff.project.domain.repositories
 import ru.itis.tinkoff.project.entity.User
 
 interface UserRepository {
-    suspend fun getUser(id: Long): User
+    suspend fun getUser(id: Int): User
 }
